Achieve image export workflow (#4)
* parallel month

* [Feat.] run through export to Drive workflow

* [Feat.] Run over the export to Drive and to Local workflow
[Feat.] Add image gennerated state record file
[Feat.] Split export task to other process to fast the workflow
[Bug fix] See ee_lst commit message
[Feat. ] Make serveral method work more faster

* remove test strction

+++ fetch_drive.py
@@ -3,7 +3,8 @@ from pydrive.drive import GoogleDrive
 import os
 import time
 from datetime import datetime
-def check_task_status(task, gap = 20):
+import logging
+def check_task_status(task, task_identifier, gap = 20):
     """
     监控任务状态直到完成或失败
     
@@ -16,26 +17,19 @@ def check_task_status(task, gap = 20):
     """
     
     while True:
-        # 获取任务状态
         status = task.status()
         state = status['state']
-        
-        # 打印当前状态
-        print(f"[{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}] 任务状态: {state}")
-        
-        # 检查是否完成
-        if state == 'COMPLETED':
-            print("✓ 任务成功完成！")
-            return True
+        if (state != 'READY'):
+            print(f"[{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}] {task_identifier} task state: {state}")
+            if state == 'COMPLETED':
+                print(f"✓ {task_identifier} task success")
+                return True
+            elif state in ['FAILED', 'CANCELLED']:
+                logging.error(f"{task_identifier} failed")
+                print(f"× {task_identifier} task failed")
+                return False
             
-        # 检查是否失败
-        elif state in ['FAILED', 'CANCELLED']:
-            error_message = status.get('error_message', '未知错误')
-            print(f"× 任务失败: {error_message}")
-            return False
-            
-        # 等待一段时间再检查
-        time.sleep(gap)  # 10秒检查一次
+        time.sleep(gap)
 
 
 def create_folder(drive,parent_folder_id,folder_name):
@@ -63,28 +57,23 @@ def get_folder_id_by_name(drive, folder_name, parent_id='root'):
         return file_list[0]['id']
     return None
 
-def download_and_clean(drive,folder_id, save_path):
-    # 创建保存目录
+def download_and_clean(drive,folder_id, cloud_file_name, save_path):
+    cloud_file_name = cloud_file_name + '.tif'
     os.makedirs(save_path, exist_ok=True)
     
-    # 查询文件夹内容（排除子文件夹）
     file_list = drive.ListFile({
         'q': f"'{folder_id}' in parents and mimeType != 'application/vnd.google-apps.folder'",
         'maxResults': 1000
     }).GetList()
     
-    # 下载并删除文件
-    for index, file_obj in enumerate(file_list):
-        try:
-            # 下载文件
-            filename = os.path.join(save_path, file_obj['title'])
-            print(f"({index+1}/{len(file_list)}) 正在下载 {file_obj['title']}")
-            file_obj.GetContentFile(filename)
-            
-            # 删除云端文件
+    for file_obj in file_list:
+        if (file_obj['title'] == cloud_file_name):
+            print(f"find file {cloud_file_name}")
+            local_file_name = os.path.join(save_path, cloud_file_name)
+            print(f"downloading to {local_file_name}")
+            file_obj.GetContentFile(local_file_name)
             file_obj.Delete()
-            print(f"√ 已成功删除云端文件")
-            
-        except Exception as e:
-            print(f"× 文件处理失败: {str(e)}")
-            continue
+            print(f"delete {cloud_file_name}")
+            return
+    print(f"not find file {cloud_file_name}")
+    return
